% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_profile.R
\name{overlap_profile}
\alias{overlap_profile}
\title{Spatial overlap profile}
\usage{
overlap_profile(raster, qvalues, metric, threshold)
}
\arguments{
\item{raster}{A RasterStack (multiple projections) object containing ENM projections with suitability scores.}

\item{qvalues}{A vector of sequential orders of diversity (default from 0 to 5). qvalues=seq(from = 0, to = 5, by = (0.1))}

\item{metric}{A vector indicating the similarity metrics to be computed. Default: metric=c("C","U","V","S")}

\item{threshold}{Suitability value(s) below which all values are converted into zeros. If a RasterStack (multiple projections) is used, the argument should contain a vector of threshold values.}
}
\value{
A vector of breadth values at different orders of diversity or a matrix containing breadth values at different orders of diversity (columns) per projection (rows).
}
\description{
Compute profiles of Hill numbers-based spatial overlaps from multiple Environmental Niche Model projection rasters.
}
\details{
Spatial overlap profile computation based on Hill numbers
}
\examples{
data()
overlap_profile(rasters)
overlap_profile(rasters,qvalues=seq(from = 0, to = 5, by = 1))
breadth_profile(rasters,qvalues=seq(from = 0, to = 5, by = 1), metric=C)
}
\references{
Alberdi, A., Novella-Fernandez R., Aizpurua, O., Razgour, O. (2019). Measuring breadth and overlap of spatial projections of environmental niche models based on Hill numbers.\cr\cr
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.\cr\cr
}
\seealso{
\code{\link{breadth}}
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{diversity}
\keyword{hill}
\keyword{overlap}
\keyword{profile}
\keyword{turnover}
